refactor: move everything into the skd directory
