+++ packages/core/index.ts
@@ -1,4 +1,2 @@
-export * from './src/@types/scenario'
-export * from './src/@types/exposition'
-export * from './src/utils'
+export * from './src/sdk'
 export * from './src/functions/createExposition'
