+++ packages/core/src/sdk/@types/scenario.ts
