+++ packages/core/src/sdk/index.ts
@@ -1,3 +1,5 @@
+export * from './@types/exposition'
+export * from './@types/scenario'
 export * from './createExpositionState'
 export * from './getExpositionValues'
 export * from './resetExpositionValues'
